made reusable Buttons and set limit 99
